Use protobuf to serialize URL (#22)
Google Flights uses protobuf to serialize the URL. Firstly, I did
serialization manually, but I think adding protobuf, even for a really
simple task, is the right approach. With this, it can be easily ported
to any language with a protocol buffer compiler.

The downside of this method is that the final binary will be ~4 MB
larger. (e.g. example3 binary 8 MB -> 12 MB).